[Task] finalize new number format function and add functionality to readme file

+++ core/locale/module.go
@@ -35,7 +35,7 @@ func (m *Module) DefaultConfig() config.Map {
 			"numbers": config.Map{
 				"decimal": ",",
 				"thousand": ".",
-				"precision": 2,
+				"precision": float64(2),
 			},
 			"date": config.Map{
 				"dateFormat":     "02 Jan 2006",
